feat: split PaymetTs&Cs and SummaryTs&Cs into files, amount calculated higher up
