+++ support-frontend/assets/pages/supporter-plus-landing/components/paymentTsAndCs.tsx
@@ -1,23 +1,11 @@
-import type { SerializedStyles } from '@emotion/react';
 import { css } from '@emotion/react';
-import {
-	neutral,
-	space,
-	textSans12,
-	textSans17,
-} from '@guardian/source/foundations';
+import { neutral, textSans12 } from '@guardian/source/foundations';
 import { StripeDisclaimer } from 'components/stripe/stripeDisclaimer';
 import type {
 	ContributionType,
 	RegularContributionType,
 } from 'helpers/contributions';
-import { formatAmount } from 'helpers/forms/checkouts';
 import type { CountryGroupId } from 'helpers/internationalisation/countryGroup';
-import type { IsoCurrency } from 'helpers/internationalisation/currency';
-import {
-	currencies,
-	spokenCurrencies,
-} from 'helpers/internationalisation/currency';
 import {
 	contributionsTermsLinks,
 	guardianAdLiteTermsLink,
@@ -31,11 +19,6 @@ import {
 	productCatalogDescription,
 } from 'helpers/productCatalog';
 import type { Promotion } from 'helpers/productPrice/promotions';
-import {
-	getDateWithOrdinal,
-	getLongMonth,
-} from 'helpers/utilities/dateFormatting';
-import type { FinePrintTheme } from './finePrint';
 import { FinePrint } from './finePrint';
 import { ManageMyAccountLink } from './manageMyAccountLink';
 
@@ -53,32 +36,6 @@ const container = css`
 		}
 	}
 `;
-
-const containerSummaryTsCs = css`
-	margin-top: ${space[6]}px;
-	border-radius: ${space[2]}px;
-	background-color: ${neutral[97]};
-	padding: ${space[3]}px;
-	${textSans17};
-	color: ${neutral[0]};
-	& a {
-		color: ${neutral[7]};
-	}
-`;
-
-interface PaymentTsAndCsProps extends SummaryTsAndCsProps {
-	countryGroupId: CountryGroupId;
-}
-
-interface SummaryTsAndCsProps {
-	mobileTheme?: FinePrintTheme;
-	contributionType: ContributionType;
-	currency: IsoCurrency;
-	amount: number;
-	productKey: ActiveProductKey;
-	promotion?: Promotion;
-	cssOverrides?: SerializedStyles;
-}
 
 const termsTierThree = (linkText: string) => (
 	<a href={tierThreeTermsLink}>{linkText}</a>
@@ -100,23 +57,6 @@ const termsLink = (linkText: string, url: string) => (
 		{linkText}
 	</a>
 );
-
-function TsAndCsRenewal({
-	contributionType,
-}: {
-	contributionType: ContributionType;
-}): JSX.Element {
-	const today = new Date();
-	if (contributionType === 'ANNUAL') {
-		return (
-			<>
-				on the {getDateWithOrdinal(today)} day of {getLongMonth(today)} every{' '}
-				year
-			</>
-		);
-	}
-	return <>on the {getDateWithOrdinal(today)} day of every month</>;
-}
 
 export function TsAndCsFooterLinks({
 	countryGroupId,
@@ -166,10 +106,18 @@ export function TsAndCsFooterLinks({
 	);
 }
 
+interface PaymentTsAndCsProps {
+	productKey: ActiveProductKey;
+	contributionType: ContributionType;
+	countryGroupId: CountryGroupId;
+	amount: number;
+	promotion?: Promotion;
+}
 export function PaymentTsAndCs({
 	productKey,
 	contributionType,
 	countryGroupId,
+	amount,
 	promotion,
 }: PaymentTsAndCsProps): JSX.Element {
 	return (
@@ -180,8 +128,7 @@ export function PaymentTsAndCs({
 					countryGroupId={countryGroupId}
 					productKey={productKey}
 					promotion={promotion}
-					currency={'GBP'}
-					amount={0}
+					amount={amount}
 				/>
 				<TsAndCsFooterLinks
 					countryGroupId={countryGroupId}
@@ -195,6 +142,7 @@ function PaymentTsAndCsComponent({
 	contributionType,
 	countryGroupId,
 	productKey,
+	amount,
 	promotion,
 }: PaymentTsAndCsProps): JSX.Element {
 	const productLabel = productCatalogDescription[productKey].label;
@@ -242,7 +190,7 @@ function PaymentTsAndCsComponent({
 						countryGroupId,
 						contributionType as RegularContributionType,
 						' per ',
-						productKey,
+						amount,
 						promotion,
 					)}
 					, you will receive the {productLabel} benefits on a subscription
@@ -287,73 +235,6 @@ function PaymentTsAndCsComponent({
 						{ManageMyAccountLink} or see our Digital + print{' '}
 						{termsLink('Terms', tierThreeTermsLink)}.
 					</p>
-				</div>
-			);
-		default:
-			return <></>;
-	}
-}
-
-export function SummaryTsAndCs({
-	productKey,
-	contributionType,
-	currency,
-	amount,
-	cssOverrides,
-}: SummaryTsAndCsProps): JSX.Element {
-	return (
-		<div css={[containerSummaryTsCs, cssOverrides]}>
-			<SummaryTsAndCsComponent
-				contributionType={contributionType}
-				currency={currency}
-				amount={amount}
-				productKey={productKey}
-			/>
-		</div>
-	);
-}
-function SummaryTsAndCsComponent({
-	contributionType,
-	currency,
-	amount,
-	productKey,
-}: SummaryTsAndCsProps): JSX.Element {
-	const amountCopy = ` of ${formatAmount(
-		currencies[currency],
-		spokenCurrencies[currency],
-		amount,
-		false,
-	)}`;
-	switch (productKey) {
-		case 'Contribution':
-			return (
-				<div>
-					We will attempt to take payment{amountCopy},{' '}
-					<TsAndCsRenewal contributionType={contributionType} />, from now until
-					you cancel your payment. Payments may take up to 6 days to be recorded
-					in your bank account. You can change how much you give or cancel your
-					payment at any time.
-				</div>
-			);
-		case 'SupporterPlus':
-			return (
-				<div>
-					The {productCatalogDescription[productKey].label} subscription and any
-					contribution will auto-renew each{' '}
-					{frequencySingular(contributionType)}. You will be charged the
-					subscription and contribution amounts using your chosen payment method
-					at each renewal, at the rate then in effect, unless you cancel.
-				</div>
-			);
-		case 'TierThree':
-		case 'GuardianAdLite':
-			return (
-				<div>
-					The {productCatalogDescription[productKey].label} subscription
-					{productKey === 'TierThree' ? 's' : ''} will auto-renew each{' '}
-					{frequencySingular(contributionType)}. You will be charged the
-					subscription amount using your chosen payment method at each renewal,
-					at the rate then in effect, unless you cancel.
 				</div>
 			);
 		default:
